Rename Bot.cs + separate logic
